+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - server: fix parsing timestamp values in BigQuery backends (fix #8076)
 - server: add support for customising the GraphQL schema descriptions of table root fields
 - console: include cron trigger with include in metadata as false on cron trigger manage page
+- console: show an error notification if Hasura CLI migrations fail
 - cli: fix metadata version being set to 3 when doing `hasura init --version 2` (#8148)
 
 ## v2.2.1